requirements.txt
torch==1.8
numpy==1.21.0
detectron2==0.4.1
matplotlib==3.4.2
pillow==8.3.1